requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# Vector store and embeddings
chromadb>=0.4.22

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0